refractoring test_application files, creating test_functions.py to hold functions
